rename file appTypes to userTypes
