+++ client/src/constants/salaryConstants.ts
@@ -2,7 +2,7 @@ import {
   KeysSalaryInit,
   SalaryInit,
 } from "../types/salaryTypes";
-import { AppLanguages } from "../types/appTypes";
+import { AppLanguages } from "../types/userTypes";
 
 const currentYear = new Date().getFullYear();
 const currentMonth = new Date().getMonth() + 1;
